# openapi.yaml
openapi: 3.1.0
info:
  title: AGP Control Plane API
  version: 0.2.0
  description: >
    OpenAPI contract derived from the authoritative AGP control-plane specs.
    This document defines the network contract for orchestration clients and runtimes.
    V1 authentication uses bearer tokens only; /health is intentionally unauthenticated.
servers:
  - url: http://127.0.0.1:7860
...
security:
  - bearerAuth: []
paths:
...
  /agents:
    get:
      tags: [agents]
      summary: List registered agents
      operationId: listAgents
      parameters:
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/AgentStatus'
        - in: query
          name: capability
          schema:
            type: string
        - in: query
          name: created_after
          schema:
            type: string
            format: date-time
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Agent list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /agents/up:
    post:
      tags: [agents]
      summary: Register ephemeral agent via self-registration
      operationId: createAgent
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentUpRequest'
      responses:
        '200':
          description: Agent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentEnvelope'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /agents/{agent_id}:
    get:
      tags: [agents]
      summary: Get agent details
      operationId: getAgent
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          description: Agent detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthenticated'
    patch:
      tags: [agents]
      summary: Update agent capabilities or metadata
      operationId: patchAgent
      parameters:
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentPatchRequest'
      responses:
        '200':
          description: Agent updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /agents/{agent_id}/interrupt:
    post:
      tags: [agents]
      summary: Request interrupt on agent's active job
      operationId: interruptAgent
      parameters:
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentInterruptRequest'
      responses:
        '200':
          description: Interrupt requested
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /agents/{agent_id}/undrain:
    post:
      tags: [agents]
      summary: Cancel drain and return agent to idle
      operationId: undrainAgent
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          description: Agent undrained
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /agents/{agent_id}/down:
    post:
      tags: [agents]
      summary: Deregister agent (drain or force-delete)
      operationId: destroyAgent
      parameters:
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentDownRequest'
      responses:
        '200':
          description: Agent teardown accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /capabilities:
    get:
      tags: [capabilities]
      summary: List capability blueprints
      operationId: listCapabilities
      parameters:
        - in: query
          name: name
          schema:
            type: string
        - in: query
          name: version
          schema:
            type: string
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Capability list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapabilityListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /capabilities/{capability_id}:
    get:
      tags: [capabilities]
      summary: Get capability blueprint
      operationId: getCapability
      parameters:
        - in: path
          name: capability_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Capability detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /capabilities/seed:
    post:
      tags: [capabilities]
      summary: Seed a capability blueprint (legacy)
      operationId: seedCapability
      deprecated: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OkEnvelope'
      responses:
        '200':
          description: Capability seeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkEnvelope'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /nudges:
    post:
      tags: [orchestration]
      summary: Create a nudge for a target agent
      operationId: createNudge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNudgeRequest'
      responses:
        '200':
          description: Nudge created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkEnvelope'
        '401':
          $ref: '#/components/responses/Unauthenticated'
    get:
      tags: [orchestration]
      summary: List nudges
      operationId: listNudges
      parameters:
        - in: query
          name: target_agent_id
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Nudge list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkEnvelope'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /nudges/next:
    get:
      tags: [orchestration]
      summary: Dequeue next pending nudge for agent
      operationId: nextNudge
      parameters:
        - in: query
          name: target_agent_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Next nudge or null
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkEnvelope'
        '401':
          $ref: '#/components/responses/Unauthenticated'
...
  /system/upgrade-status:
    get:
      tags: [health]
      summary: Get persisted release and schema version state
      operationId: getUpgradeStatus
      responses:
        '200':
          description: Upgrade status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpgradeStatusEnvelope'
        '401':
          $ref: '#/components/responses/Unauthenticated'
...
components:
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: false
      schema:
        type: string
    Cursor:
      in: query
      name: cursor
      schema:
        type: string
    Limit:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 100
...
    AgentId:
      in: path
      name: agent_id
      required: true
      schema:
        type: string
...
  responses:
    InvalidRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthenticated:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    ErrorCode:
      type: string
      enum:
        - invalid_request
        - not_found
        - conflict
        - stale_fencing_token
        - lease_expired
        - already_terminal
        - not_eligible
        - unauthenticated
        - forbidden
        - rate_limited
        - internal_error
...
    AgentStatus:
      type: string
      enum: [idle, busy, draining]
...
    ErrorEnvelope:
      type: object
      required: [ok, error]
      properties:
        ok:
          type: boolean
          const: false
        error:
          type: object
          required: [code, message, retryable]
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message:
              type: string
            retryable:
              type: boolean
    PageInfo:
      type: object
      required: [has_more]
      properties:
        next_cursor:
          type: string
        has_more:
          type: boolean
...
    AgentSummary:
      type: object
      required: [agent_id, status, queue_id]
      properties:
        agent_id:
          type: string
        capabilities:
          type: array
          items:
            type: string
        metadata_json:
          type: object
        queue_id:
          type: string
        status:
          $ref: '#/components/schemas/AgentStatus'
        workspace_ref:
          type: string
        last_heartbeat_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
...
    CapabilitySummary:
      type: object
      required: [capability_id, name, version, image_ref, model_ref, resource_tier, permission_profile, queue_mode]
      properties:
        capability_id:
          type: string
        name:
          type: string
        version:
          type: string
        image_ref:
          type: string
        model_ref:
          type: string
        resource_tier:
          type: string
        permission_profile:
          type: string
        queue_mode:
          type: string
        runtime_requirements_json:
          type: object
          additionalProperties: true
...
    AgentPatchRequest:
      type: object
      properties:
        capabilities:
          type: array
          items:
            type: string
        metadata:
          type: object
          additionalProperties: true
        workspace_ref:
          type: string
    AgentInterruptRequest:
      type: object
      required: [fencing_token]
      properties:
        fencing_token:
          type: integer
          format: int64
        reason:
          type: string
    CreateNudgeRequest:
      type: object
      required: [target_agent_id]
      properties:
        target_agent_id:
          type: string
        priority:
          type: integer
        source:
          type: string
        payload:
          type: object
          additionalProperties: true
        job_id:
          type: string
    AgentUpRequest:
      type: object
      properties:
        agent_id:
          type: string
        capabilities:
          type: array
          items:
            type: string
        metadata:
          type: object
        workspace_ref:
          type: string
    AgentDownRequest:
      type: object
      properties:
        mode:
          type: string
          enum: [drain, force]
    AgentEnvelope:
      type: object
      required: [ok, data]
      properties:
        ok:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/AgentSummary'
    AgentListEnvelope:
      type: object
      required: [ok, data]
      properties:
        ok:
          type: boolean
          const: true
        data:
          type: object
          required: [items, page]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/AgentSummary'
            page:
              $ref: '#/components/schemas/PageInfo'
    CapabilityListEnvelope:
      type: object
      required: [ok, data]
      properties:
        ok:
          type: boolean
          const: true
        data:
          type: object
          required: [items, page]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/CapabilitySummary'
            page:
              $ref: '#/components/schemas/PageInfo'
...
    UpgradeStatus:
      type: object
      required: [release_version, schema_version, package_version]
      properties:
        release_version:
          type: string
        schema_version:
          type: string
        previous_release_version:
          type: string
        previous_schema_version:
          type: string
        package_version:
          type: string
        rollback_target_release_version:
          type: string
        rollback_target_schema_version:
          type: string
    UpgradeStatusEnvelope:
      type: object
      required: [ok, data]
      properties:
        ok:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/UpgradeStatus'
...
    OkEnvelope:
      type: object
      required: [ok, data]
      properties:
        ok:
          type: boolean
          const: true
        data:
          type: object
          additionalProperties: true
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque